requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
chardet==3.0.4
click==7.1.2
colorama==0.3.7
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.1.4
deluge==1.3.15
distlib==0.3.1
enum34==1.1.6
filelock==3.0.12
Flask==1.1.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==2.1.1
importlib-resources==3.3.1
incremental==16.10.1
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
numpy==1.13.3
onetimepad==1.4
PAM==0.4.2
pathlib2==2.3.5
pdfshuffler==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pygame===1.9.1release
pygobject==3.26.1
pyOpenSSL==17.5.0
PyPDF2==1.26.0
pyserial==3.4
python-libtorrent==1.1.5
pyxdg==0.25
scandir==1.10.0
SecretStorage==2.3.1
service-identity==16.0.0
singledispatch==3.6.1
six==1.15.0
termcolor==1.1.0
Twisted==17.9.0
typing==3.7.4.3
virtualenv==20.4.2
Werkzeug==1.0.1
zipp==1.2.0
zope.interface==4.3.2
